encode integers in the smallest BSON type possible
For ordinary Perl integers, they are encoded as BSON Int32 if they fit
and BSON Int64 otherwise.

For Math::BigInt objects, conversion now extracts the BigInt as a string
and process with strtoll.  The previous implementation violated
encapsulation and made assumptions about the inner structure of
Math::BigInt objects.  Then, the result will be encoded as BSON Int32 if
it fits and BSON Int64 otherwise.